Using native int instead of intXX_t types

+++ libraspifb16/image565Qoi.cxx
@@ -259,8 +259,8 @@ decodeQoi(
             }
         }
 
-        auto x = static_cast<int16_t>(i % header.getWidth());
-        auto y = static_cast<int16_t>(i / header.getWidth());
+        auto x = i % header.getWidth();
+        auto y = i / header.getWidth();
 
         raspifb16::RGB565 rgb{currentRGBA.r, currentRGBA.g, currentRGBA.b};
         image.setPixelRGB(raspifb16::Image565Point{x, y}, rgb);
